fix int64 issue; fix random seed issue

+++ subdivnet/mesh_tensor.py
@@ -289,9 +289,9 @@ class MeshTensor:
                 self.faces[i, :F, [1, 2]],
                 self.faces[i, :F, [2, 0]],
                 self.faces[i, :F, [0, 1]],
-            ], dim=0).astype('int64')
+            ], dim=0)
 
-            E_hash = E.min(dim=1) * E.max() + E.max(dim=1)
+            E_hash = E.min(dim=1).astype('int64') * E.max() + E.max(dim=1)
 
             # S is index of sorted E_hash.
             # Based on the construction rule of E,
